ParkingGarage.java
A parking garage has 5 customers daily. The garage charges a $5.00 minimum fee to park up to two hours. The garage charges an additional $1.00 per hour for each hour (or part of an hour) over two hours.  The maximum charge for any given day is $12.00.  All cars are gone by midnight.  Write a program to calculate and print a summary of the charges for a day.  For input the program will read the hours of usage for each of the 5 cars.  The program will print the result, as shown below.